Added Split function for the Toom-Cook part
Successfully tested Split and then Solve with the output matrix using 16-bit numbers with k=3 and points 0, infinity, 1, -1, 2. Infinity and 0 are always chosen by default because they always lead to small computations. Negative integer points seem to be twice as efficient as their positive counterparts. For example evaluating at -2 leads to working with numbers twice smaller than the numbers we would end up making the algorithm work with if we evaluate at 2. The points for input:
1000000000000000
1111111111111111
0100000000000000
The only thing left to do is to code up the final recomposition part, which is to start at the top result in the final matrix and then start at the second-lowest result and going up, right-shifting by X each time, where X is how many digits each split number can have at most, until we reach the second highest, then we go back to the bottom result for the final shift and we add all these numbers together to produce the final result of the multiplication, which in this test case is indeed 106,284,230, or as my algorithm correctly worked it out: 0110010101011100010011000110, where the leftmost bit is the least significant bit.